+++ examples/stallion.py
@@ -99,7 +99,6 @@ logger = TensorBoardLogger(log_graph=True)
 trainer = pl.Trainer(
     max_epochs=100,
     gpus=0,
-    weights_summary="top",
     gradient_clip_val=0.1,
     limit_train_batches=30,
     # val_check_interval=20,
@@ -133,7 +132,7 @@ print(f"Number of parameters in network: {tft.size()/1e3:.1f}k")
 # trainer.limit_train_batches = 1.0
 # # run learning rate finder
 # res = trainer.tuner.lr_find(
-#     tft, train_dataloader=train_dataloader, val_dataloaders=val_dataloader, min_lr=1e-5, max_lr=1e2
+#     tft, train_dataloaders=train_dataloader, val_dataloaders=val_dataloader, min_lr=1e-5, max_lr=1e2
 # )
 # print(f"suggested learning rate: {res.suggestion()}")
 # fig = res.plot(show=True, suggest=True)
@@ -142,7 +141,7 @@ print(f"Number of parameters in network: {tft.size()/1e3:.1f}k")
 
 # trainer.fit(
 #     tft,
-#     train_dataloader=train_dataloader,
+#     train_dataloaders=train_dataloader,
 #     val_dataloaders=val_dataloader,
 # )
 
@@ -178,6 +177,6 @@ with open("test_study.pkl", "wb") as fout:
 #     model=tft,
 #     period=0.001,
 #     filter="pytorch_forecasting",
-#     train_dataloader=train_dataloader,
+#     train_dataloaders=train_dataloader,
 #     val_dataloaders=val_dataloader,
 # )