+++ pytorch_forecasting/models/base_model.py
@@ -12,6 +12,7 @@ import numpy as np
 from numpy.lib.function_base import iterable
 import pandas as pd
 from pytorch_lightning import LightningModule
+from pytorch_lightning.trainer.states import RunningStage
 from pytorch_lightning.utilities.parsing import AttributeDict, get_init_args
 import scipy.stats
 import torch
@@ -33,6 +34,7 @@ from pytorch_forecasting.metrics import (
     MultiHorizonMetric,
     MultiLoss,
     QuantileLoss,
+    convert_torchmetric_to_pytorch_forecasting_metric,
 )
 from pytorch_forecasting.optim import Ranger
 from pytorch_forecasting.utils import (
@@ -118,6 +120,15 @@ def _concatenate_output(
     if isinstance(output[0], OutputMixIn):
         output_cat = output[0].__class__(**output_cat)
     return output_cat
+
+
+STAGE_STATES = {
+    RunningStage.TRAINING: "train",
+    RunningStage.VALIDATING: "val",
+    RunningStage.TESTING: "test",
+    RunningStage.PREDICTING: "predict",
+    RunningStage.SANITY_CHECKING: "sanity_check",
+}
 
 
 class BaseModel(LightningModule):
@@ -234,11 +245,13 @@ class BaseModel(LightningModule):
 
         if not hasattr(self, "loss"):
             if isinstance(loss, (tuple, list)):
-                self.loss = MultiLoss(metrics=loss)
+                self.loss = MultiLoss(metrics=[convert_torchmetric_to_pytorch_forecasting_metric(l) for l in loss])
             else:
-                self.loss = loss
+                self.loss = convert_torchmetric_to_pytorch_forecasting_metric(loss)
         if not hasattr(self, "logging_metrics"):
-            self.logging_metrics = nn.ModuleList([l for l in logging_metrics])
+            self.logging_metrics = nn.ModuleList(
+                [convert_torchmetric_to_pytorch_forecasting_metric(l) for l in logging_metrics]
+            )
         if not hasattr(self, "output_transformer"):
             self.output_transformer = output_transformer
         if not hasattr(self, "optimizer"):  # callables are removed from hyperparameters, so better to save them
@@ -260,6 +273,14 @@ class BaseModel(LightningModule):
         for k in hparams_to_delete:
             del self._hparams[k]
             del self._hparams_initial[k]
+
+    @property
+    def current_stage(self) -> str:
+        """
+        Available inside lightning loops.
+        :return: current trainer stage. One of ["train", "val", "test", "predict", "sanity_check"]
+        """
+        return STAGE_STATES[self.trainer.state.stage]
 
     @property
     def n_targets(self) -> int:
@@ -371,6 +392,18 @@ class BaseModel(LightningModule):
         log.update(self.create_log(x, y, out, batch_idx))
         return log
 
+    def validation_epoch_end(self, outputs):
+        self.epoch_end(outputs)
+
+    def test_step(self, batch, batch_idx):
+        x, y = batch
+        log, out = self.step(x, y, batch_idx)
+        log.update(self.create_log(x, y, out, batch_idx))
+        return log
+
+    def test_epoch_end(self, outputs):
+        self.epoch_end(outputs)
+
     def create_log(
         self,
         x: Dict[str, torch.Tensor],
@@ -403,9 +436,6 @@ class BaseModel(LightningModule):
                 x, out, batch_idx, prediction_kwargs=prediction_kwargs, quantiles_kwargs=quantiles_kwargs
             )
         return {}
-
-    def validation_epoch_end(self, outputs):
-        self.epoch_end(outputs)
 
     def step(
         self, x: Dict[str, torch.Tensor], y: Tuple[torch.Tensor, torch.Tensor], batch_idx: int, **kwargs
@@ -500,7 +530,14 @@ class BaseModel(LightningModule):
             else:
                 loss = self.loss(prediction, y)
 
-        self.log(f"{['val', 'train'][self.training]}_loss", loss, on_step=self.training, on_epoch=True, prog_bar=True)
+        self.log(
+            f"{self.current_stage}_loss",
+            loss,
+            on_step=self.training,
+            on_epoch=True,
+            prog_bar=True,
+            batch_size=len(x["decoder_target"]),
+        )
         log = {"loss": loss, "n_samples": x["decoder_lengths"].size(0)}
         return log, out
 
@@ -548,10 +585,11 @@ class BaseModel(LightningModule):
                 else:
                     target_tag = ""
                 self.log(
-                    f"{target_tag}{['val', 'train'][self.training]}_{metric.name}",
+                    f"{target_tag}{self.current_stage}_{metric.name}",
                     loss_value,
                     on_step=self.training,
                     on_epoch=True,
+                    batch_size=len(x["decoder_target"]),
                 )
 
     def to_network_output(self, **results):
@@ -662,7 +700,7 @@ class BaseModel(LightningModule):
                 log_indices = [0]
             for idx in log_indices:
                 fig = self.plot_prediction(x, out, idx=idx, add_loss_to_title=True, **kwargs)
-                tag = f"{['Val', 'Train'][self.training]} prediction"
+                tag = f"{self.current_stage} prediction"
                 if self.training:
                     tag += f" of item {idx} in global batch {self.global_step}"
                 else:
@@ -875,11 +913,15 @@ class BaseModel(LightningModule):
             except TypeError:  # in case there is no weight decay
                 optimizer = self.optimizer(self.parameters(), lr=lr, **optimizer_params)
         elif self.hparams.optimizer == "adam":
-            optimizer = torch.optim.Adam(self.parameters(), lr=lr)
+            optimizer = torch.optim.Adam(
+                self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay, **optimizer_params
+            )
         elif self.hparams.optimizer == "adamw":
-            optimizer = torch.optim.AdamW(self.parameters(), lr=lr)
+            optimizer = torch.optim.AdamW(
+                self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay, **optimizer_params
+            )
         elif self.hparams.optimizer == "ranger":
-            optimizer = Ranger(self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay)
+            optimizer = Ranger(self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay, **optimizer_params)
         elif self.hparams.optimizer == "sgd":
             optimizer = torch.optim.SGD(
                 self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay, **optimizer_params
@@ -898,34 +940,31 @@ class BaseModel(LightningModule):
         if isinstance(lr, (list, tuple)):  # change for each epoch
             # normalize lrs
             lrs = np.array(lrs) / lrs[0]
-            schedulers = [
-                {
-                    "scheduler": LambdaLR(optimizer, lambda epoch: lrs[min(epoch, len(lrs) - 1)]),
-                    "interval": "epoch",
-                    "reduce_on_plateau": False,
-                    "frequency": 1,
-                }
-            ]
-        elif self.val_dataloader() is None:  # no schedule and no validation loss means no scheduling to be used
-            schedulers = []
+            scheduler_config = {
+                "scheduler": LambdaLR(optimizer, lambda epoch: lrs[min(epoch, len(lrs) - 1)]),
+                "interval": "epoch",
+                "frequency": 1,
+                "strict": False,
+            }
+        elif self.hparams.reduce_on_plateau_patience is None:
+            scheduler_config = {}
         else:  # find schedule based on validation loss
-            schedulers = [
-                {
-                    "scheduler": ReduceLROnPlateau(
-                        optimizer,
-                        mode="min",
-                        factor=0.1,
-                        patience=self.hparams.reduce_on_plateau_patience,
-                        cooldown=self.hparams.reduce_on_plateau_patience,
-                        min_lr=self.hparams.reduce_on_plateau_min_lr,
-                    ),
-                    "monitor": "val_loss",  # Default: val_loss
-                    "interval": "epoch",
-                    "reduce_on_plateau": True,
-                    "frequency": 1,
-                }
-            ]
-        return [optimizer], schedulers
+            scheduler_config = {
+                "scheduler": ReduceLROnPlateau(
+                    optimizer,
+                    mode="min",
+                    factor=0.2,
+                    patience=self.hparams.reduce_on_plateau_patience,
+                    cooldown=self.hparams.reduce_on_plateau_patience,
+                    min_lr=self.hparams.reduce_on_plateau_min_lr,
+                ),
+                "monitor": "val_loss",  # Default: val_loss
+                "interval": "epoch",
+                "frequency": 1,
+                "strict": False,
+            }
+
+        return {"optimizer": optimizer, "lr_scheduler": scheduler_config}
 
     @classmethod
     def from_dataset(cls, dataset: TimeSeriesDataSet, **kwargs) -> LightningModule: