+++ pytorch_forecasting/models/basic_rnn/__init__.py
@@ -100,7 +100,7 @@ class LSTMModel(AutoRegressiveBaseModelWithCovariates):
 
         time_series_rnn = get_rnn(cell_type)
         cont_size = len(self.reals)
-        cat_size = sum([size[1] for size in self.hparams.embedding_sizes.values()])
+        cat_size = sum(self.embeddings.output_size.values())
         input_size = cont_size + cat_size
         self.rnn = time_series_rnn(
             input_size=input_size,