+++ pytorch_forecasting/models/temporal_fusion_transformer/tuning.py
@@ -143,7 +143,7 @@ def optimize_hyperparameters(
                 PyTorchLightningPruningCallback(trial, monitor="val_loss"),
             ],
             logger=logger,
-            progress_bar_refresh_rate=[0, 1][optuna_verbose < optuna.logging.INFO],
+            enable_progress_bar=optuna_verbose < optuna.logging.INFO,
             weights_summary=[None, "top"][optuna_verbose < optuna.logging.INFO],
         )
         default_trainer_kwargs.update(trainer_kwargs)
@@ -174,12 +174,12 @@ def optimize_hyperparameters(
                 gradient_clip_val=gradient_clip_val,
                 gpus=[0] if torch.cuda.is_available() else None,
                 logger=False,
-                progress_bar_refresh_rate=0,
-                weights_summary=None,
+                enable_progress_bar=False,
+                enable_model_summary=False,
             )
             res = lr_trainer.tuner.lr_find(
                 model,
-                train_dataloader=train_dataloader,
+                train_dataloaders=train_dataloader,
                 val_dataloaders=val_dataloader,
                 early_stop_threshold=10000,
                 min_lr=learning_rate_range[0],
@@ -206,7 +206,7 @@ def optimize_hyperparameters(
             model.hparams.learning_rate = trial.suggest_loguniform("learning_rate", *learning_rate_range)
 
         # fit
-        trainer.fit(model, train_dataloader=train_dataloader, val_dataloaders=val_dataloader)
+        trainer.fit(model, train_dataloaders=train_dataloader, val_dataloaders=val_dataloader)
 
         # report result
         return metrics_callback.metrics[-1]["val_loss"].item()