+++ tests/test_data/test_timeseries.py
@@ -1,7 +1,9 @@
 from copy import deepcopy
+import itertools
 import pickle
 from typing import Dict
 
+import networkx
 import numpy as np
 import pandas as pd
 import pytest
@@ -53,9 +55,95 @@ def test_categorical_target(test_data):
         min_prediction_length=1,
         min_encoder_length=1,
     )
-
     _, y = next(iter(dataset.to_dataloader()))
     assert y[0].dtype is torch.long, "target must be of type long"
+
+
+@pytest.mark.parametrize(
+    "kwargs",
+    [
+        dict(method="robust"),
+        dict(method="robust", data=np.random.randn(100)),
+        dict(data=np.random.randn(100)),
+        dict(transformation="log"),
+        dict(transformation="softplus"),
+        dict(transformation="log1p"),
+        dict(transformation="relu"),
+        dict(method="identity"),
+        dict(method="identity", data=np.random.randn(100)),
+        dict(center=False),
+        dict(max_length=5),
+        dict(data=pd.Series(np.random.randn(100))),
+        dict(max_length=[1, 2]),
+    ],
+)
+def test_EncoderNormalizer(kwargs):
+    kwargs.setdefault("method", "standard")
+    kwargs.setdefault("center", True)
+    kwargs.setdefault("data", torch.rand(100))
+    data = kwargs.pop("data")
+
+    normalizer = EncoderNormalizer(**kwargs)
+    if kwargs.get("transformation") in ["relu", "softplus"]:
+        data = data - 0.5
+
+    if kwargs.get("transformation") in ["relu", "softplus", "log1p"]:
+        assert (
+            normalizer.inverse_transform(torch.as_tensor(normalizer.fit_transform(data))) >= 0
+        ).all(), "Inverse transform should yield only positive values"
+    else:
+        assert torch.isclose(
+            normalizer.inverse_transform(torch.as_tensor(normalizer.fit_transform(data))),
+            torch.as_tensor(data),
+            atol=1e-5,
+        ).all(), "Inverse transform should reverse transform"
+
+
+def test_EncoderNormalizer_with_limited_history():
+    data = torch.rand(100)
+    normalizer = EncoderNormalizer(max_length=[1, 2]).fit(data)
+    assert normalizer.center_ == data[-1]
+
+
+@pytest.mark.parametrize(
+    "kwargs,groups",
+    itertools.product(
+        [
+            dict(method="robust"),
+            dict(transformation="log"),
+            dict(transformation="relu"),
+            dict(center=False),
+            dict(transformation="log1p"),
+            dict(transformation="softplus"),
+            dict(scale_by_group=True),
+        ],
+        [[], ["a"]],
+    ),
+)
+def test_GroupNormalizer(kwargs, groups):
+    data = pd.DataFrame(dict(a=[1, 1, 2, 2, 3], b=[1.1, 1.1, 1.0, 5.0, 1.1]))
+    defaults = dict(method="standard", transformation=None, center=True, scale_by_group=False)
+    defaults.update(kwargs)
+    kwargs = defaults
+    kwargs["groups"] = groups
+    kwargs["scale_by_group"] = kwargs["scale_by_group"] and len(kwargs["groups"]) > 0
+
+    if kwargs.get("transformation") in ["relu", "softplus"]:
+        data.b = data.b - 2.0
+    normalizer = GroupNormalizer(**kwargs)
+    encoded = normalizer.fit_transform(data["b"], data)
+
+    test_data = dict(
+        prediction=torch.tensor([encoded[0]]),
+        target_scale=torch.tensor(normalizer.get_parameters([1])).unsqueeze(0),
+    )
+
+    if kwargs.get("transformation") in ["relu", "softplus", "log1p"]:
+        assert (normalizer(test_data) >= 0).all(), "Inverse transform should yield only positive values"
+    else:
+        assert torch.isclose(
+            normalizer(test_data), torch.tensor(data.b.iloc[0]), atol=1e-5
+        ).all(), "Inverse transform should reverse transform"
 
 
 def test_pickle(test_dataset):
@@ -525,3 +613,11 @@ def test_graph_sampler(test_dataset):
         if idx > 100:
             break
     print(a)
+
+
+def test_TorchNormalizer_dtype_consistency():
+    """Ensures that even for float64 `target_scale`, the transformation will not change the prediction dtype."""
+    parameters = torch.tensor([[[366.4587]]])
+    target_scale = torch.tensor([[427875.7500, 80367.4766]], dtype=torch.float64)
+    assert TorchNormalizer()(dict(prediction=parameters, target_scale=target_scale)).dtype == torch.float32
+    assert TorchNormalizer().transform(parameters, target_scale=target_scale).dtype == torch.float32