+++ tests/test_models/test_deepar.py
@@ -20,6 +20,7 @@ from pytorch_forecasting.models import DeepAR
 def _integration(
     data_with_covariates, tmp_path, gpus, cell_type="LSTM", data_loader_kwargs={}, clip_target: bool = False, **kwargs
 ):
+    data_with_covariates = data_with_covariates.copy()
     if clip_target:
         data_with_covariates["target"] = data_with_covariates["volume"].clip(1e-3, 1.0)
     else:
@@ -33,26 +34,30 @@ def _integration(
     )
     data_loader_default_kwargs.update(data_loader_kwargs)
     dataloaders_with_covariates = make_dataloaders(data_with_covariates, **data_loader_default_kwargs)
+
     train_dataloader = dataloaders_with_covariates["train"]
     val_dataloader = dataloaders_with_covariates["val"]
+    test_dataloader = dataloaders_with_covariates["test"]
+
     early_stop_callback = EarlyStopping(monitor="val_loss", min_delta=1e-4, patience=1, verbose=False, mode="min")
 
     logger = TensorBoardLogger(tmp_path)
     trainer = pl.Trainer(
         max_epochs=3,
         gpus=gpus,
-        weights_summary="top",
         gradient_clip_val=0.1,
         callbacks=[early_stop_callback],
-        checkpoint_callback=True,
+        enable_checkpointing=True,
         default_root_dir=tmp_path,
         limit_train_batches=2,
         limit_val_batches=2,
+        limit_test_batches=2,
         logger=logger,
     )
 
     net = DeepAR.from_dataset(
         train_dataloader.dataset,
+        hidden_size=5,
         cell_type=cell_type,
         learning_rate=0.15,
         log_gradient_flow=True,
@@ -64,9 +69,11 @@ def _integration(
     try:
         trainer.fit(
             net,
-            train_dataloader=train_dataloader,
+            train_dataloaders=train_dataloader,
             val_dataloaders=val_dataloader,
         )
+        test_outputs = trainer.test(net, dataloaders=test_dataloader)
+        assert len(test_outputs) > 0
         # check loading
         net = DeepAR.load_from_checkpoint(trainer.checkpoint_callback.best_model_path)
 
@@ -102,7 +109,7 @@ def _integration(
         ),
         dict(
             data_loader_kwargs=dict(
-                lags={"volume": [2, 5]}, target="volume", time_varying_unknown_reals=["volume"], min_encoder_length=10
+                lags={"volume": [2, 5]}, target="volume", time_varying_unknown_reals=["volume"], min_encoder_length=2
             )
         ),
         dict(
@@ -125,6 +132,7 @@ def model(dataloaders_with_covariates):
     dataset = dataloaders_with_covariates["train"].dataset
     net = DeepAR.from_dataset(
         dataset,
+        hidden_size=5,
         learning_rate=0.15,
         log_gradient_flow=True,
         log_interval=1000,